pyproject.toml
[tool.poetry]
name = "pyawscron"
version = "1.0.7"
description = "An AWS CRON Expression Parser"
authors = ["Michael Martin <pitchblack408@gmail.com>", "Steven Gorrell <stevengor@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.9.0.post0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"
flake8-use-fstring = "^1.4"
flake8-print = "^5.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
force-exclude = '''
(
  /(
    \.eggs
    | \.git
    | \.demo
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | out
    | docs
    | dependencies
  )
)
'''
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
skip = [
  ".demo",
  ".eggs",
  ".git",
  ".mypy_cache",
  ".venv",
  "_build",
  "build",
  "dist",
  "out",
  "docs",
  "venv",
]
